+++ src/keybind.cpp
@@ -959,7 +959,7 @@ void	kf_ExpandScreen( void )
 /* Spins the world round left */
 void	kf_RotateLeft()
 {
-	float rotAmount = realTimeAdjustedIncrement(MAP_SPIN_RATE);
+	int rotAmount = static_cast<int>(realTimeAdjustedIncrement(MAP_SPIN_RATE));
 
 	playerPos.r.y += rotAmount;
 }
@@ -968,7 +968,7 @@ void	kf_RotateLeft()
 /* Spins the world right */
 void	kf_RotateRight()
 {
-	float rotAmount = realTimeAdjustedIncrement(MAP_SPIN_RATE);
+	int rotAmount = static_cast<int>(realTimeAdjustedIncrement(MAP_SPIN_RATE));
 
 	playerPos.r.y -= rotAmount;
 	if (playerPos.r.y < 0)
@@ -995,7 +995,7 @@ void kf_RotateBuildingACW()
 /* Pitches camera back */
 void	kf_PitchBack()
 {
-	float pitchAmount = realTimeAdjustedIncrement(MAP_PITCH_RATE);
+	int pitchAmount = static_cast<int>(realTimeAdjustedIncrement(MAP_PITCH_RATE));
 
 	playerPos.r.x += pitchAmount;
 
@@ -1009,7 +1009,7 @@ void	kf_PitchBack()
 /* Pitches camera forward */
 void	kf_PitchForward()
 {
-	float pitchAmount = realTimeAdjustedIncrement(MAP_PITCH_RATE);
+	int pitchAmount = static_cast<int>(realTimeAdjustedIncrement(MAP_PITCH_RATE));
 
 	playerPos.r.x -= pitchAmount;
 	if (playerPos.r.x < DEG(360 + MIN_PLAYER_X_ANGLE))
